+++ src/ExtendedCharTable.h
@@ -29,7 +29,7 @@ namespace Konsole {
 /**
  * A table which stores sequences of unicode characters, referenced
  * by hash keys.  The hash key itself is the same size as a unicode
- * character ( ushort ) so that it can occupy the same space in
+ * character ( uint ) so that it can occupy the same space in
  * a structure.
  */
 class ExtendedCharTable
@@ -50,7 +50,7 @@ public:
      * @param unicodePoints An array of unicode character points
      * @param length Length of @p unicodePoints
      */
-    ushort createExtendedChar(const ushort *unicodePoints, ushort length);
+    uint createExtendedChar(const uint *unicodePoints, ushort length);
     /**
      * Looks up and returns a pointer to a sequence of unicode characters
      * which was added to the table using createExtendedChar().
@@ -61,20 +61,20 @@ public:
      *
      * @return A unicode character sequence of size @p length.
      */
-    ushort *lookupExtendedChar(ushort hash, ushort &length) const;
+    uint *lookupExtendedChar(uint hash, ushort &length) const;
 
     /** The global ExtendedCharTable instance. */
     static ExtendedCharTable instance;
 private:
     // calculates the hash key of a sequence of unicode points of size 'length'
-    ushort extendedCharHash(const ushort *unicodePoints, ushort length) const;
+    uint extendedCharHash(const uint *unicodePoints, ushort length) const;
     // tests whether the entry in the table specified by 'hash' matches the
     // character sequence 'unicodePoints' of size 'length'
-    bool extendedCharMatch(ushort hash, const ushort *unicodePoints, ushort length) const;
-    // internal, maps hash keys to character sequence buffers.  The first ushort
-    // in each value is the length of the buffer, followed by the ushorts in the buffer
+    bool extendedCharMatch(uint hash, const uint *unicodePoints, ushort length) const;
+    // internal, maps hash keys to character sequence buffers.  The first uint
+    // in each value is the length of the buffer, followed by the uints in the buffer
     // themselves.
-    QHash<ushort, ushort *> _extendedCharTable;
+    QHash<uint, uint *> _extendedCharTable;
 };
 }
 #endif  // end of EXTENDEDCHARTABLE_H
